+++ documentation/ChangeLog.md
@@ -5,6 +5,7 @@
  #TODO
  * Added syntax coloring support for `?` operator.
  * Invoking a build operation for a project now immediately cancels the pending build operation for that project, if any.
+ * More lenient with `rustfmt` exit code 3, don't treat those warnings as a failure. (#144)
 
 ### 0.7.0
  * Added support for Rust 1.12 new error message format. (#130)